% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart_stations}
\alias{bart_stations}
\title{List all BART stations}
\usage{
bart_stations()
}
\value{
tibble: abbr, name, city, county, state, latitude, longitude,
address, zipcode
}
\description{
List all BART stations
}
